waltio: move waltio code to its own file
